+++ ext_test.go
@@ -4,8 +4,8 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/gsamokovarov/msgpack"
-	"github.com/gsamokovarov/msgpack/codes"
+	"github.com/mgechev/msgpack"
+	"github.com/mgechev/msgpack/codes"
 )
 
 func init() {